Bring query modal into its own component. Filter is not working
