+++ frontend/pages/hosts/HostDetailsPage/SelectQueryModal/index.js
@@ -1,0 +1,1 @@
+export { default } from "./SelectQueryModal";
